// packages/vitest/_src/app/model/SpecificationContext.ts
/**
 * Framework metadata about the specification
 * READ-ONLY - contains file/title/description/tags
 * NOT for user test data! Use local variables instead.
 */
export class SpecificationContext {
    filename: string = "";
    title: string = "";
    description: string = "";
    tags: string[] = [];
}